+++ packages/RSuite/NEWS.md
@@ -1,4 +1,4 @@
-# RSuite 0.36 (2019-03-??)
+# RSuite 0.36 (2019-03-06)
   * #184: detecting declaration of dependencies in scripts with "library (...)".
   * handling of branches in Git RC adapter fixed.
   * #183: local Git repository is created while starting project if not RC detected